requirements.txt
alembic==1.14.1
Flask==3.1.0
Flask_Bootstrap==3.3.7.1
Flask_Login==0.6.3
flask_mail==0.10.0
Flask_Migrate==4.1.0
Flask_Moment==1.0.6
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
itsdangerous==2.2.0
python-dotenv==1.1.0
SQLAlchemy==2.0.38
Werkzeug==3.1.3
WTForms==3.2.1
